requirements.txt
# Radix Sylva — aligné Jardin bIOT (Django 5.2)
Django==5.2.11
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
drf-spectacular==0.28.0
Pillow==11.0.0
requests==2.32.3
psycopg[binary]>=3.2
